Handle window resize

+++ src/OrionEngine/Orion/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -2,6 +2,7 @@
 #include "OpenGLVertexArray.h"
 
 #include <glad/glad.h>
+#include <cstdint>
 
 namespace Orion
 {
@@ -66,7 +67,7 @@ namespace Orion
                                   ShaderDataTypeToOpenGLBaseType(element.Type),
                                   element.Normalized ? GL_TRUE : GL_FALSE,
                                   static_cast<GLsizei>(layout.GetStride()),
-                                  reinterpret_cast<const void*>(element.Offset));
+                                  reinterpret_cast<const void*>(static_cast<std::uintptr_t>(element.Offset)));
             index++;
         }
